Take access to RequestSchema away from the dataset
