+++ snuba/datasets/__init__.py
@@ -2,8 +2,6 @@ from typing import Optional, Mapping
 
 from snuba.datasets.dataset_schemas import DatasetSchemas
 from snuba.util import escape_col
-from snuba.schemas import RequestSchema, GENERIC_QUERY_SCHEMA
-from snuba.query.query_processor import ExtensionQueryProcessor
 from snuba.query.extensions import QueryExtension
 
 
@@ -81,7 +79,7 @@ class Dataset(object):
         """
         raise NotImplementedError
 
-    def _get_extensions(self) -> Mapping[str, QueryExtension]:
+    def get_extensions(self) -> Mapping[str, QueryExtension]:
         """
         Returns the extensions for this dataset.
         Every extension comes as a tuple (schema, processor).
@@ -90,25 +88,6 @@ class Dataset(object):
         extension.
         """
         raise NotImplementedError('dataset does not support queries')
-
-    def get_query_schema(self) -> RequestSchema:
-        generic_schema = GENERIC_QUERY_SCHEMA
-        extensions_schemas = {
-            extension_key: extension.get_schema()
-            for extension_key, extension
-            in self._get_extensions().items()
-        }
-        return RequestSchema(
-            generic_schema,
-            extensions_schemas,
-        )
-
-    def get_query_processors(self) -> Mapping[str, ExtensionQueryProcessor]:
-        return {
-            extension_key: extension.get_processor()
-            for extension_key, extension
-            in self._get_extensions().items()
-        }
 
 
 class TimeSeriesDataset(Dataset):
